fargate2/requirements.txt
configparser
mysql.connector
boto3
sentry-sdk